sky130_ef_io*.lib: operating_conditions() use a quoted-string
Add surrounding quotemarks (example):
  operating_conditions (deep_sleep_lfclk_max_-40) {
Transformed into:
  operating_conditions ("deep_sleep_lfclk_max_-40") {

The use of the minus character in something that would otherwise look like an
identifer might be problematic for some parsers.  Everything should be happy
if it was quoted as the specification allows.

Liberty_Reference_Manual_R-2020-09.pdf page 72 describes the group type as
'name-string' as opposed to 'name-id'.